fix: move analytics helpers out of resolvers file to prevent go generate from breaking them
